Added render pass timestamp profiling.
* Added description for /GL command.
* Switched to auto references for lambdas.
* Added GPU timestamps for render passes. If enabled, they can be retrieved after the frame renders via a call to: QueryTimeStamps in TinyRenderPass.

Timestamps are tracked internally each frame by writing timestamps to CommandBuffers that are recorded to. An internal counter must be kept to track timestamp writes, because VkQueryPool requires us to manually reset individual timestamp queries instead of resetting the whole query pool...

Each timestamp requires 2 queries one at the top-of-pipe and one at the bottom-of-pipe in order to profile the entire VkQueueSubmit pipeline. Internally 2x queries are tracked and only half (the deltas) are returned by QueryTimeStamps in TinyRenderPass.

NOTE: Timestamps are AUTOMATICALLY disabled when building for release.